requirements.txt
flask
gunicorn
scikit-learn
pandas
numpy
PyMuPDF
python-docx
